% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeSeriesAnalysis.R
\docType{package}
\name{TimeSeriesAnalysis-package}
\alias{TimeSeriesAnalysis}
\alias{TimeSeriesAnalysis-package}
\title{TimeSeriesAnalysis: Time series analysis exploration for cohorts generated against the OMOP Common Data Model}
\description{
Time series analysis exploration for cohorts generated against the OMOP Common Data Model
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/EHDEN/TimeSeriesAnalysis}
  \item Report bugs at \url{https://github.com/EHDEN/TimeSeriesAnalysis/issues}
}

}
\author{
\strong{Maintainer}: Anthony Sena \email{asena5@its.jnj.com}

Authors:
\itemize{
  \item Sara Khalid
}

}
\keyword{internal}
